requirements.txt
MetaTrader5; platform_system == "Windows"
flask==3.0.3
gunicorn==22.0.0
sqlalchemy==2.0.32
psycopg[binary]==3.2.13
requests==2.32.3
PyJWT==2.8.0
